Apollo-Backend
About Apollo

Apollo is a creative mobile application designed to provide an interactive and user-friendly interface for users to listen to various songs, especially Rwandan songs. apollo will also allow the users to do challenges of the different songs they like and share them in one application without having to navigate through different applications. Sharing songs and music is also convenient as one can post them on their story easily.

Note: This is a backend repository to handle backend codes in Nodejs/Express

Planning and Development Process

The system is intended to be developed in at least 6 months with basic and core functionalities including:
•	Landing page with interactive features
•	Registration page
•	Preference page
•	Dashboard with songs of different chosen genres with a search field
•	Playlist page
•	Current playing song and its details
•	Challenge page 
•	Notification page
•	Profile management page includes followers, following, posts, saved posts, highlights
•	Create a new challenge page
•	Story page and add reaction feature
•	Notifications from comments and any other updates

NB: - In order to complete all the tasks, we need to be updated at least twice a week.
       -  We need to cooperate to make it easier and faster